# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Uúba Tech API",
...
    "contact": {
      "name": "Uúba Tech",
      "url": "https://uuba.tech/"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/clientes": {
      "post": {
        "tags": [
          "clientes"
        ],
        "summary": "Cadastrar cliente",
        "description": "Registra um novo cliente na carteira. O campo `documento` (CPF ou CNPJ) deve ser único.",
        "operationId": "create_cliente_api_v1_clientes_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "clientes"
        ],
        "summary": "Listar clientes",
        "description": "Retorna a lista de clientes com suporte a filtro por telefone. Use `telefone` para buscar pelo número WhatsApp.",
        "operationId": "list_clientes_api_v1_clientes_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "telefone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por número WhatsApp (ex: 5511999001234)",
              "title": "Telefone"
            },
            "description": "Filtrar por número WhatsApp (ex: 5511999001234)"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ordenação (ex: total_vencido)",
              "title": "Order By"
            },
            "description": "Ordenação (ex: total_vencido)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "description": "Itens por página (max 100)",
              "default": 50,
              "title": "Limit"
            },
            "description": "Itens por página (max 100)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pular N itens",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pular N itens"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes/{cliente_id}": {
      "get": {
        "tags": [
          "clientes"
        ],
        "summary": "Buscar cliente",
        "description": "Retorna os dados de um cliente específico pelo ID.",
        "operationId": "get_cliente_api_v1_clientes__cliente_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cliente Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "clientes"
        ],
        "summary": "Atualizar cliente",
        "description": "Atualiza parcialmente os dados de um cliente. Envie apenas os campos que deseja alterar.",
        "operationId": "update_cliente_api_v1_clientes__cliente_id__patch",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cliente Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes/{cliente_id}/metricas": {
      "get": {
        "tags": [
          "clientes"
        ],
        "summary": "Métricas de pagamento",
        "description": "Retorna indicadores financeiros do cliente: DSO (dias médios para pagamento), total em aberto, total vencido, e contagem de faturas.",
        "operationId": "get_metricas_api_v1_clientes__cliente_id__metricas_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cliente Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteMetricas"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClienteCreate": {
        "properties": {
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "documento": {
            "type": "string",
            "title": "Documento"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "documento"
        ],
        "title": "ClienteCreate"
      },
      "ClienteMetricas": {
        "properties": {
          "dso_dias": {
            "type": "number",
            "title": "Dso Dias",
            "default": 0.0
          },
          "total_em_aberto": {
            "type": "integer",
            "title": "Total Em Aberto",
            "default": 0
          },
          "total_vencido": {
            "type": "integer",
            "title": "Total Vencido",
            "default": 0
          },
          "faturas_em_aberto": {
            "type": "integer",
            "title": "Faturas Em Aberto",
            "default": 0
          },
          "faturas_vencidas": {
            "type": "integer",
            "title": "Faturas Vencidas",
            "default": 0
          }
        },
        "type": "object",
        "title": "ClienteMetricas"
      },
      "ClienteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "cliente"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "documento": {
            "type": "string",
            "title": "Documento"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "documento",
          "created_at",
          "updated_at"
        ],
        "title": "ClienteResponse"
      },
      "ClienteUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          }
        },
        "type": "object",
        "title": "ClienteUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListResponse": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object",
            "default": "list"
          },
          "data": {
            "items": {},
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "ListResponse"
      },
      "PaginationMeta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "total",
          "page_size",
          "has_more"
        ],
        "title": "PaginationMeta"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}